src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appdirs==1.4.4            # via black
asgiref==3.3.1            # via django
attrs==20.3.0             # via flake8-bugbear, flake8-eradicate, pytest
awesome-slugify==1.6.5    # via -r requirements.in
backcall==0.2.0           # via ipython
black==20.8b1             # via flake8-black
certifi==2020.12.5        # via requests, sentry-sdk
chardet==3.0.4            # via requests
click==7.1.2              # via black, pip-tools
decorator==4.4.2          # via ipython
django-behaviors==0.5.1   # via -r requirements.in
django-cors-headers==3.6.0  # via -r requirements.in
django-environ==0.4.5     # via -r requirements.in
django-healthchecks==1.4.2  # via -r requirements.in
django==3.1.4             # via -r requirements.in, django-cors-headers, django-healthchecks, djangorestframework, djangorestframework-simplejwt, drf-nested-routers
djangorestframework-simplejwt==4.6.0  # via -r requirements.in
djangorestframework==3.12.2  # via -r requirements.in, djangorestframework-simplejwt, drf-nested-routers
drf-nested-routers==0.92.5  # via -r requirements.in
environ==1.0              # via -r requirements.in
eradicate==2.0.0          # via flake8-eradicate
faker==0.9.1              # via mixer
flake8-black==0.2.1       # via -r requirements.in
flake8-bugbear==20.11.1   # via -r requirements.in
flake8-eradicate==1.0.0   # via -r requirements.in
flake8-isort==4.0.0       # via -r requirements.in
flake8-print==4.0.0       # via -r requirements.in
flake8==3.8.4             # via -r requirements.in, flake8-black, flake8-bugbear, flake8-eradicate, flake8-isort, flake8-print
freezegun==1.0.0          # via -r requirements.in, pytest-freezegun
idna==2.10                # via requests
iniconfig==1.1.1          # via pytest
ipython-genutils==0.2.0   # via traitlets
ipython==7.19.0           # via -r requirements.in
isort==5.6.4              # via -r requirements.in, flake8-isort
jedi==0.17.2              # via ipython
mccabe==0.6.1             # via flake8
mixer==6.1.3              # via -r requirements.in
mypy-extensions==0.4.3    # via black
packaging==20.8           # via pytest
parso==0.7.1              # via jedi
pathspec==0.8.1           # via black
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==5.4.0          # via -r requirements.in
pluggy==0.13.1            # via pytest
prompt-toolkit==3.0.8     # via ipython
psycopg2-binary==2.8.6    # via -r requirements.in
ptyprocess==0.6.0         # via pexpect
py==1.10.0                # via pytest
pycodestyle==2.6.0        # via flake8, flake8-print
pyflakes==2.2.0           # via flake8
pygments==2.7.3           # via ipython
pyjwt==1.7.1              # via djangorestframework-simplejwt
pyparsing==2.4.7          # via packaging
pytest-django==4.1.0      # via -r requirements.in
pytest-env==0.6.2         # via -r requirements.in
pytest-freezegun==0.4.2   # via -r requirements.in
pytest-mock==3.4.0        # via -r requirements.in
pytest==6.2.1             # via -r requirements.in, pytest-django, pytest-env, pytest-freezegun, pytest-mock
python-dateutil==2.8.1    # via faker, freezegun
pytz==2020.4              # via django
regex==2020.11.13         # via awesome-slugify, black
requests==2.25.0          # via django-healthchecks
sentry-sdk==0.19.5        # via -r requirements.in
six==1.15.0               # via django-healthchecks, faker, flake8-print, pip-tools, python-dateutil
sqlparse==0.4.1           # via django
testfixtures==6.17.0      # via flake8-isort
text-unidecode==1.2       # via faker
toml==0.10.2              # via black, pytest
traitlets==5.0.5          # via ipython
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
unidecode==0.4.21         # via awesome-slugify
urllib3==1.26.2           # via requests, sentry-sdk
wcwidth==0.2.5            # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
